requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai>=1.35.0
httpx==0.27.0
httpcore==1.0.5
python-dotenv==1.0.0
gunicorn==21.2.0
